requirements.txt
Flask==2.0.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
mysql-connector-python==8.3.0
gunicorn==21.2.0
